==> src/0.json_to_csv.ipynb <==
## JSON to CSV

import json
from os import listdir
from os.path import isfile, join
import pandas as pd

path = 'data/1.raw_csv/'

def jsontodf(dir):
  # files: 디렉토리 안의 파일 목록의 리스트
  files = [f for f in listdir(dir) if isfile(join(dir, f))]

  # json_files: 파일 목록 DataFrame
  json_files = pd.DataFrame({"file_id" : files})

  # json file 목록을 읽으며
  # 각 파일의 json 형태로 저장된 txt 파일을 DataFrame으로 형변환 한 후
  # concat를 통해 전체 데이터프레임 형성
  df_list = []
  for i in range(len(json_files)-1):
      with open(dir+json_files['file_id'][i], encoding = 'utf-8') as f:
          df = pd.DataFrame(json.loads(line) for line in f)
      df_list.append(df)
  dfs = pd.concat(df_list)

  return dfs

### log

log = jsontodf('data/0.raw_json/m3/m3_click_stream/')

log.to_csv(path + 'log.csv', encoding = 'utf-8-sig', index=None)

### Accounts

accounts = jsontodf('data/0.raw_json/m3/m3_accounts/')

accounts.to_csv(path + 'accounts.csv', encoding = 'utf-8-sig', index=None)

### Products

products = jsontodf('data/0.raw_json/m3/all_products/')

products.to_csv(path + 'products.csv', encoding = 'utf-8-sig', index=None)

### Orders

orders = jsontodf('data/0.raw_json/m3/m3_orders/')

orders.to_csv(path + 'orders.csv', encoding = 'utf-8-sig', index=None)
