==> src/1.preprocessing.ipynb <==
import pandas as pd
import numpy as np

import datetime
from dateutil.parser import parse
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import LabelEncoder
ohe = OneHotEncoder()
le = LabelEncoder()

raw_path = '../data/1.raw_csv/'
path = '../data/2.Preprocessed/'

## Accounts

accounts = pd.read_csv(raw_path + 'accounts.csv')

#### accounts: 최근 접속 일자 변환

accounts['last_login_dts'] = pd.to_datetime(accounts['last_login_dts']*1000000)
accounts.dropna(inplace=True)

#### accounts: gender 변환

accounts.gender.unique()

accounts['gender'].replace(to_replace = '-', value = np.nan, inplace = True)
accounts['gender'].replace(to_replace = '0', value = np.nan, inplace = True)
accounts.dropna(inplace=True)
accounts = pd.get_dummies(accounts, columns=['gender'])

#### accounts: age 결측처리

accounts.age.unique()

accounts['age'].replace(to_replace = -1, value = np.nan, inplace = True) # 결측값 = -1
accounts['age'] = accounts['age'].apply(lambda x : np.nan if x <=7 else x) # 7세 이하는 결측처리

accounts.dropna(inplace=True)
accounts['age'].isnull().sum()

#### accounts: age 이산화

accounts['age'] = pd.cut(accounts['age'], [8, 12, 19, 30, 42, 51, 64, np.inf],
                        labels=['0', '1', '2', '3', '4', '5', '6'], right=False)

accounts = pd.get_dummies(accounts, columns=['age'])

#### accounts: 주소

accounts['address'][0].split(' ')[0]

# 수도권만 1, 나머지 0으로 설정
accounts['address_is'] = accounts['address'].apply(lambda x: 0 if '경기' in x.split(' ')[0] else 1  if '인천' in x.split(' ')[0] else 1 if '서울' in x.split(' ')[0] else 1)

accounts.dropna(inplace=True)
accounts['address_is'].isnull().sum()

accounts = pd.get_dummies(accounts, columns=['address_is'])

gen_col = [col for col in accounts.columns if col.startswith('gender')]
add_col = [col for col in accounts.columns if col.startswith('address_is')]
age_col = [col for col in accounts.columns if col.startswith('age')]

#### 저장

accounts = accounts[['account_id', 'last_login_dts'] + gen_col + add_col + age_col]
accounts.to_csv(path + 'accounts_preprocessed.csv',encoding='utf-8', index=None)

## Orders

### orders: java시간 변환

orders = pd.read_csv(raw_path + 'orders.csv')

def invert_javatime(javatime):
  seconds = javatime / 1000
  sub_seconds = (javatime % 1000.0) / 1000.0
  date = datetime.datetime.fromtimestamp(seconds + sub_seconds)
  return date

java_timestamp = 1589424016190
invert_javatime(java_timestamp)

orders['created_at'] = orders['created_at'].apply(lambda x : invert_javatime(x))
orders

### 저장

orders.to_csv(path+'orders_preprocessed.csv',encoding='utf-8')

## Products

products = pd.read_csv(raw_path+'products.csv',dtype={'published_at':str})

products = products[products['published_at'].notna()]

### products: 출판일 전처리(date형식)

products['published_at'] = products['published_at'].apply(lambda x : x[0:8] if type(x) == 'str' else x)
products['published_at'] = products['published_at'].apply(lambda x : x.strip() if type(x) == 'str' else x)
products['published_at'] = products['published_at'].apply(lambda x : '20991231' if len(x) <8 else x)
products['published_at'] = products['published_at'].astype(str)
products['published_at'] = products['published_at'].apply(lambda x : '20991231' if int(x[4:6]) >= 13 else x )
products['published_at'] = products['published_at'].apply(lambda x : '20991231' if int(x) <= 19000000 else x)

products['published_at'] = products['published_at'].apply(lambda x : parse(x) )

products = products[products['published_at'].isin(pd.date_range('1800-01-01', '2020-12-31'))]

### products: 출판일 이산화

default_date = '2020-05-01'
def date_range(days=None, default_date='2020-05-01'):
    default_date = pd.to_datetime(default_date)
    delta = datetime.timedelta(days=days)
    before_date = default_date - delta
    after_date = default_date + delta
    return pd.date_range(before_date, after_date, periods=2)

products['pub'] = 4
products.loc[products[products['published_at'].isin(date_range(days=1095))].index,'pub'] = 3
products.loc[products[products['published_at'].isin(date_range(days=365))].index,'pub'] = 2
products.loc[products[products['published_at'].isin(date_range(days=180))].index,'pub'] = 1
products.loc[products[products['published_at'].isin(date_range(days=90))].index,'pub'] = 0

products = pd.get_dummies(products, columns=['pub'])

### products: 카테고리ID 전처리

products['category_id'] = products['category_id'].astype(str)

products['category_id'] = products['category_id'].apply(lambda x: x[2:4] if len(x) > 3 else (x+((4-len(x))*'0'))[2:4])

products = products[products['category_id'] != '00']
products = pd.get_dummies(products, columns=['category_id'])

### products: 가격대 이상치 처리 및 이산화

condition1 = products['shop_price']>3500000.0 # 3500000 만원 이상은 고서적, 일반인들이 구매하지 않을 종류의 책들, 300000~3500000 가격의 책들은 대부분 전집에 해당
condition2 = products['shop_price']<1000.0 # 1000 미만은 사은품 등의 다른 제품들이 혼재

products = products[~condition1 & ~condition2]
products = products[products['shop_price'] != 0.0]

products['shop_price'] = pd.qcut(products['shop_price'], 4, labels=['0', '1', '2', '3'])

products = pd.get_dummies(products, columns=['shop_price'])

products.to_csv(path+'products_preprocessed.csv',encoding='utf-8', index=None)

## click

click = pd.read_csv(raw_path+'log.csv')

print(click.shape)
click.head()

### click: device_type 원핫인코딩

click = pd.get_dummies(click, columns=['device_type'])

click.isnull().sum()

click.to_csv(path+'click.csv',encoding='utf-8', index=None)

## click & orders

### click & orders: orders 중 click없이 구매한 데이터(외부 사이트 이용) 삭제하기
* click 데이터 크기가 방대하여 절반만 사용한 관계로 더 적어짐

orders = pd.read_csv(path+'orders_preprocessed.csv',encoding='utf-8')

order_check_account = click['account_id'].unique().tolist()
order_check_product = click['product_id'].unique().tolist()

orders.shape

order_check_1 = orders[orders['product_id'].isin(order_check_product)]

order_check_1.shape

order_check_2 = order_check_1[order_check_1['account_id'].isin(order_check_account)]

order_check_2.shape

order_check_2 = order_check_2.iloc[:,1:]

order_check_2.to_csv(path+'orders_only_in_clicks.csv',encoding='utf-8', index=None)
