==> src/3.MergeData.ipynb <==
import pandas as pd

pre_path = '../data/2.Preprocessed/'
old_path = '../data/3.Feature_Engineered/'
new_path = '../data/4.Merged/'

# 유저

users = pd.read_csv(pre_path + 'accounts_preprocessed.csv')
users

userdf_mergelist = []

## 신구 선호도

### 주문 기반 신구 선호도

orders_newpref= pd.read_csv(old_path+'user_orders_new_preference.csv')

### 클릭 기반 신구 선호도

clicks_newpref = pd.read_csv(old_path+'user_click_new_preference.csv')

newpref = orders_newpref.copy()
newpref['new_preference'] = (0.7*orders_newpref['new_preference']) + (0.3 *clicks_newpref['new_preference'])

newpref

userdf_mergelist.append(newpref)

## 활동시간

user_click_acttime = pd.read_csv(old_path + 'user_click_activitytimes.csv')
userdf_mergelist.append(user_click_acttime)

## 카테고리 선호도

user_cate_pref = pd.read_csv(old_path + 'user_click_cat_preference.csv')
userdf_mergelist.append(user_cate_pref)

## 관여도

user_inv = pd.read_csv(old_path + 'user_involvement.csv')
userdf_mergelist.append(user_inv)

for df in userdf_mergelist:
    users = pd.merge(users, df, how='left', on='account_id')

users.to_csv(new_path + 'fed_users.csv', index=None)

# 책상품

products = pd.read_csv(pre_path+'products_preprocessed.csv')
products

## bestseller

pro_best = pd.read_csv(old_path + 'bestseller_product.csv')

pro_best

products = pd.merge(products, pro_best, how='left', on='product_id')

products.to_csv(new_path + 'fed_products.csv', index=None)

products
