==> src/5.MAB.ipynb <==
import pandas as pd
import numpy as np
import time
sample_path = '../data/5.sampled/'
k_mean_path = '../data/6.K-mean/'
mab_path = '../data/7.MAB/'

from bandits.core.contextual_bandit import run_contextual_bandit
from bandits.algorithms.linear_full_posterior_sampling import LinearFullPosteriorSampling
from bandits.data.data_sampler import  remove_underrepresented_classes

#결과출력함
def display_results(algos, opt_rewards, opt_actions, h_rewards, t_init, name):
    """Displays summary statistics of the performance of each algorithm."""

    print('---------------------------------------------------')
    print('---------------------------------------------------')
    print('{} bandit completed after {} seconds.'.format(
        name, time.time() - t_init))
    print('---------------------------------------------------')

    performance_pairs = []
    for j, a in enumerate(algos):
        performance_pairs.append((a.name, np.sum(h_rewards[:, j])))
    performance_pairs = sorted(performance_pairs,
                             key=lambda elt: elt[1],
                             reverse=True)
    for i, (name, reward) in enumerate(performance_pairs):
        print('{:3}) {:20}| \t \t total reward = {:10}.'.format(i, name, reward))

    print('---------------------------------------------------')
    print('Optimal total reward = {}.'.format(np.sum(opt_rewards)))
    print('Frequency of optimal actions (action, frequency):')
    print([[elt, list(opt_actions).count(elt)] for elt in set(opt_actions)])
    print('---------------------------------------------------')
    print('---------------------------------------------------')

 def clf_to_bandit_problem(contexts, labels, num_actions=None):
  """Normalize contexts and encode deterministic rewards."""

  if num_actions is None:
    num_actions = np.max(labels) + 1
  num_contexts = contexts.shape[0]

  # Due to random subsampling in small problems, some features may be constant
  #sstd = safe_std(np.std(contexts, axis=0, keepdims=True)[0, :])

  # Normalize features
  #contexts = ((contexts - np.mean(contexts, axis=0, keepdims=True)) / sstd)
  # One hot encode labels as rewards
  rewards = np.zeros((num_contexts, num_actions))
  rewards[np.arange(num_contexts), labels] = 1.0

  return contexts, rewards, (np.ones(num_contexts), labels)

sample_clicks = pd.read_csv(sample_path+'sampled_clicks.csv')
sample_orders = pd.read_csv(sample_path+'sampled_orders.csv')

merged_log = pd.merge(sample_clicks, sample_orders, how='left', on=['account_id', 'product_id'])

merged_log['reward'] = -1
merged_log.loc[merged_log[merged_log['order_id'].isna()].index ,'reward'] = 0
merged_log.loc[merged_log[merged_log['order_id'].notna()].index , 'reward'] = 1

interaction = merged_log[['account_id', 'product_id', 'reward', 'request_date_time']]
interaction.to_csv(mab_path+'interaction.csv', index=None)

sample_products = pd.read_csv(k_mean_path+'kmeans_product_onehotcol.csv')
sample_users = pd.read_csv(k_mean_path+'kmeans_user_onehotcol.csv')

user_inter = pd.merge(interaction, sample_users, how='left', on='account_id')
user_pro_inter = pd.merge(user_inter, sample_products, how='left', on='product_id')

user_pro_inter.drop(columns=['last_login_dts', 'product_name', 'published_at', 'maker_name'], inplace=True)
user_pro_inter.rename({'cluster_x': 'user_cluster', 'cluster_y' : 'book_cluster'}, axis=1, inplace=True)

user_pro_inter.rename(lambda x:'most_pref' + '_' + x.split('_')[-2] if x.endswith('x') else x,  axis=1, inplace=True)
user_pro_inter.rename(lambda x:x[:-2] if x.endswith('y') else x,  axis=1, inplace=True)

user_pro_inter.insert(2, 'user__cluster', user_pro_inter['user_cluster'])
user_pro_inter.insert(3, 'book__cluster', user_pro_inter['book_cluster'])
user_pro_inter.drop(['user_cluster','book_cluster'], axis=1, inplace=True)

user_pro_inter.to_csv(mab_path + 'sampled_mabdata.csv', index=None)

df_fin = pd.read_csv(mab_path + 'sampled_mabdata.csv')
df_fin.dropna(inplace=True)


df_fin.sort_values('request_date_time', inplace=True)

df_fin

(context, reward_1, )

opt_actions

dataset = np.hstack((contexts, rewards))
context_dim = contexts.shape[1]

class Hparams(object):
    def __init__(self, dicts):
        for k, v in dicts.items():
            self.__setattr__(k, v)

param_dicts = {'num_actions':num_actions,
 'context_dim': context_dim,
 'a0':6,
 'b0':6,
 'lambda_prior':0.25,
  'initial_pulls':2}

hparam = Hparams(param_dicts)

algos = [LinearFullPosteriorSampling('LinFullPost', hparam)]

t_init = time.time()
results = run_contextual_bandit(context_dim, num_actions, dataset, algos)

h_actions, h_rewards = results

display_results(algos, opt_rewards, opt_actions, h_rewards, t_init, 'context')
